Standardize Readme
This is the most minimal change to standardize the README, following RichardLitt/standard-readme. See #558 for more on this.

Big changes: Move the title and the badges, take away the centering. This is a bit less pretty, but you should not have to think about this stuff, and standardizing will remove some aesthetic decisions, sadly. I also added a contribute, a Table of Contents, and moved badges into usage - I think this is the best move. Linked to the license. What do you think?