Add a disambiguating "standardjs" bin alias
Because there are now "standard"-branded opinionated linters being 
written for other languages, it seems smart to give people who use both
a way to disambiguate between the two by running `standardrb`, 
`standardjs`, etc. explicitly.

Someday, it'd be really neat if a meta `standard` bin could roll-up 
linting for multiple languages in a single project in a single
invocation, but for now I think this is a decent way to remediate the
risk of people globally installing multiple standard packages and 
invoking each from the user PATH.