Update package.json
Remove "funding" adware.

"funding" is a "paused" experiment and this commit removes it as a dependency from standard/standard, least ads come back in build logs unexpectedly.  By leaving this dependency in place it allows feross/funding to turn ads back on in standard/standard without any changes in standard/standard, this commit prevents that.